modificar metodos repository

+++ src/main/java/com/logistica/logistica_principal/service/PedidoService.java
@@ -1,7 +1,6 @@
 package com.logistica.logistica_principal.service;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,25 +19,23 @@ public class PedidoService {
         return pedidorepository.findAll();
     }
 
-    public Optional<PedidoEntity> buscarPorId(Integer idPedido){
-        return pedidorepository.findById(idPedido);
+    public PedidoEntity buscarPorId(Integer idPedido){
+        if (pedidorepository.existsById(idPedido)) {
+            return pedidorepository.findPedidoByIdPedido(idPedido);
+        } 
+        return null; 
     }
 
-    // se cambia pedido a pedidodto
-    private PedidoDto convertirDto(PedidoEntity entity){
-        PedidoDto dto = new PedidoDto();
-        dto.setComunaPedido(entity.getComunaPedido());
-        dto.setEstadoPedido(entity.getEstadoPedido());
-        dto.setFechaEntrega(entity.getFechaEntrega());
-        return dto;
-    }
-
-    public List<PedidoDto> buscarPorEstado(String estadoPedido){
+    public List<PedidoDto> buscarPorEstado(String estado){
         try {
-            List<PedidoEntity> pedidos = pedidorepository.findByEstadoPedido(estadoPedido);
+            List<PedidoEntity>pedidos = pedidorepository.findByEstadoPedido(estado);
             List<PedidoDto> pedidoDtos = new ArrayList<>();
-            for (PedidoEntity pedido : pedidos){
-                pedidoDtos.add(convertirDto(pedido));
+            for (PedidoEntity pedido : pedidos) {
+                pedidoDtos.add(new PedidoDto(
+                    pedido.getComunaPedido(),
+                    pedido.getEstadoPedido(),
+                    pedido.getFechaEntrega()
+                ));  
             }
             return pedidoDtos;
         } catch (Exception e) {
@@ -46,55 +43,48 @@ public class PedidoService {
         }
     }
 
-
-    public List<PedidoDto> buscarPorComuna(String comunaPedido){
+    public  String agregarPedido(PedidoEntity pedido){
         try {
-            List<PedidoEntity> pedidos = pedidorepository.findByComunaPedido(comunaPedido);
-            List<PedidoDto> pedidoDtos = new ArrayList<>();
-            for (PedidoEntity pedido : pedidos) {
-                pedidoDtos.add(convertirDto(pedido)); 
-            }
-            return pedidoDtos;
+            PedidoEntity pedidonuevo = new PedidoEntity();
+            pedidonuevo.setComunaPedido(pedido.getComunaPedido());
+            pedidonuevo.setFechaCompra(pedido.getFechaCompra());
+            pedidonuevo.setFechaEntrega(pedido.getFechaEntrega());
+            pedidonuevo.setEstadoPedido(pedido.getEstadoPedido());
+            pedidorepository.save(pedidonuevo);
+            return "Pedido agregado correctamente";
         } catch (Exception e) {
-            return new ArrayList<>();  //si da error, devuelve una lista vacia
-        } 
-    }
-    
-
-    public PedidoEntity agregarPedido(PedidoEntity pedido){
-        PedidoEntity pedidonuevo = new PedidoEntity();
-        pedidonuevo.setComunaPedido(pedido.getComunaPedido());
-        pedidonuevo.setFechaCompra(pedido.getFechaCompra());
-        pedidonuevo.setFechaEntrega(pedido.getFechaEntrega());
-        pedidonuevo.setEstadoPedido(pedido.getEstadoPedido());
-        return pedidorepository.save(pedidonuevo);
+            return "Error al agregar el pedido: "+ e.getMessage();
+        }
     }
 
     //elimina pedido por id
-    public void eliminarPedido(Integer idPedido){
-        pedidorepository.deleteById(idPedido);
+    public String eliminarPedido(Integer idPedido){
+        if (pedidorepository.existsById(idPedido)) {
+            pedidorepository.deleteById(idPedido);
+            return "Pedido eliminado correctamente";
+        }
+        return "Pedido no encontrado";
     }
 
     //actualiza un pedido existente
-    public Optional<PedidoEntity> actualizaPedido(PedidoEntity pedidoActualizado){
+    public String actualizaPedido(PedidoEntity pedidoActualizado){
         try {
             Integer idPedido = pedidoActualizado.getIdPedido();
-            Optional<PedidoEntity> pedidoExiste = pedidorepository.findById(idPedido);
             
-            if (pedidoExiste.isPresent()) {
-                PedidoEntity pedido = pedidoExiste.get();
+            if (pedidorepository.existsById(idPedido)) {
+                PedidoEntity pedido = pedidorepository.findPedidoByIdPedido(idPedido);
                 pedido.setComunaPedido(pedidoActualizado.getComunaPedido());
                 pedido.setFechaCompra(pedidoActualizado.getFechaCompra());
                 pedido.setFechaEntrega(pedidoActualizado.getFechaEntrega());
                 pedido.setEstadoPedido(pedidoActualizado.getEstadoPedido());
-                PedidoEntity pedidoGuardado = pedidorepository.save(pedido);
-                return Optional.of(pedidoGuardado);
+                pedidorepository.save(pedido);
+                return "Pedido actualizado";
             }else{
-                return Optional.empty();
+                return "Pedido no encontrado";
             }
             
         } catch (Exception e) {
-            return Optional.empty();
+            return "Error al actualizar pedido: "+ e.getMessage();
         }
     }
